+++ src/task-edit.js
@@ -1,19 +1,11 @@
-class TaskEdit {
-  constructor(data) {
-    this._title = data.title;
-    this._dueDate = data.dueDate;
-    this._tags = data.tags;
-    this._image = data.image;
-    this._color = data.color;
-    this._repeatingDays = data.repeatingDays;
-    this._element = null;
+import {TaskComponent} from "./taskComponent";
 
-    this._state = {
-      isFavorite: data.isFavorite,
-      isDone: data.isDone,
-    };
+class TaskEdit extends TaskComponent {
+  constructor(data) {
+    super(data);
 
     this._onSubmit = null;
+    this._onSubmitButtonClick = this._onSubmitButtonClick.bind(this);
   }
 
   /**
@@ -47,66 +39,6 @@ class TaskEdit {
       </label>
     </div>
   `;
-  }
-
-  /**
-   * Возвращает класс для смены класса кнопок задачи
-   * @param {boolean} isDisabled
-   * @return {string}
-   */
-  _getTaskDisabledClass(isDisabled) {
-    const TaskDisabledClassList = {
-      true: ``,
-      false: ` card__btn--disabled`
-    };
-
-    return TaskDisabledClassList[isDisabled];
-  }
-
-  /**
-   * Возвращает является ли задача просроченной
-   * @param {*} dueDateStr
-   * @return {boolean}
-   */
-  _getIsDeadline(dueDateStr) {
-    const curDate = new Date();
-    const dueDateArr = dueDateStr.split(` `);
-    const dueDate = new Date(Date.parse(`${dueDateArr[1]} ${dueDateArr[0]}, 2019`));
-    return curDate > dueDate;
-  }
-
-  /**
-   * Возвращает цвет бара задачи в зависимости её просроченности
-   * @param {string} color
-   * @param {boolean} isDeadline
-   * @return {string}
-   */
-  _getTaskColor(color, isDeadline) {
-    return isDeadline ? `deadline` : color;
-  }
-
-  /**
-   * Возвращает повторяется задача или нет
-   * @param {object} repeatingDays
-   * @return {boolean}
-   */
-  _getIsRepeatTask(repeatingDays) {
-    for (let dayValue in repeatingDays) {
-      if (repeatingDays[dayValue]) {
-        return dayValue;
-      }
-    }
-
-    return false;
-  }
-
-  /**
-   * Возвращает класс повторяющейся задачи
-   * @param {boolean} isRepeat
-   * @return {string}
-   */
-  _getTaskBarType(isRepeat) {
-    return isRepeat ? `card--repeat` : ``;
   }
 
   /**
@@ -155,7 +87,7 @@ class TaskEdit {
    * Отрисовка карточки задачи
    * @return {string}
    */
-  get taskEditTemplate() {
+  get taskTemplate() {
     const taskControlBlock = `
     <div class="card__control">
       <button type="button" class="card__btn card__btn--edit card__btn--disabled">
@@ -269,10 +201,6 @@ class TaskEdit {
     }
   }
 
-  get element() {
-    return this._element;
-  }
-
   set onSubmit(fn) {
     this._onSubmit = fn;
   }
@@ -288,25 +216,6 @@ class TaskEdit {
   removeEvents() {
     const cardBtnSubmitElem = this._element.querySelector(`.card__save`);
     cardBtnSubmitElem.removeEventListener(`click`, this._onSubmitButtonClick.bind(this));
-  }
-
-  /**
-   * Отрисовка задачи в заданном элементе
-   * @return {link}
-   */
-  render() {
-    this._element = null || document.createElement(`div`);
-
-    this._element.innerHTML = this.taskEditTemplate;
-    this._element = this._element.firstElementChild;
-    this.addEvents();
-
-    return this._element;
-  }
-
-  unrender() {
-    this.removeEvents();
-    this._element = null;
   }
 }
